+++ src/madflight/common/MF_Serial.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 // Serial Interface
 class MF_Serial {
   public: